+++ Assets/Scripts/InventoryManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44960aed1ee480d43b32876f96cc26ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
