+++ Assets/Scripts/TextManager.cs
