+++ Assets/Scripts/TextManager.cs.meta
