+++ Assets/Sprites/wall-removebg-preview.png.meta
